codermates-test-series/test-backend/partial_question_management/serializers.py
from rest_framework import serializers
from question_management.models import Question

class QuestionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Question
        fields = '__all__'
